Cleaned out old/out-of-date/unnecessary stuff, recreated wordlists with proper documentation.
Details:
Deleted all the django/database-related stuff.
Deleted all previously-existing wordlists.
Deleted json_tools.py and longitudinal.py as being out-of-date and unnecessary.
Modified __init__ to remove references to django/database-related stuff.
Added corpora_used.txt, which is a reference list of all the NLTK corpora used to construct the wordlists.
Added words.txt, which is a list of all the words extracted from the corpora cited in corpora_used.txt.
Added words_lowered.txt, which is the same words as in words.txt, but lowercased using .lower().
Added freqs.json, which is a json dump of a dictionary of words from words.txt and their frequencies in the list.
Added lowered_freqs.json, which is a json dump of a dictionary of words from lowered_words.txt and their frequencies in the list.